pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mdlearn"
version = "1.0.2"
authors = [
    {name = "Alexander Brace", email = "abrace@anl.gov"},
    {name = "Heng Ma", email = "heng.ma@anl.gov"},
    {name = "Debsindhu Bhowmik", email = "bhowmikd@ornl.gov"},
    {name = "Chakra Chennubhotla", email = "chakracs@pitt.edu"},
    {name = "Arvind Ramanathan", email = "ramanathana@anl.gov"}
]
description = "Machine learning for molecular dynamics."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "pyyaml",
    "pydantic",
    "h5py",
    "pandas",
    "numpy",
    "tqdm",
    "plotly",
]

[project.urls]
homepage = "https://github.com/ramanathanlab/mdlearn"
documentation = "https://mdlearn.readthedocs.io/en/latest"
repository = "https://github.com/ramanathanlab/mdlearn"

[project.optional-dependencies]
dev = [
    "covdefaults>=2.2",
    "coverage",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff",
    "tox",
    "virtualenv",
    "types-PyYAML",
    "build",
    "twine",
]
docs = [
    "black",
    "sphinx==8.1.3",
    "sphinx-rtd-theme==3.0.1",
    "autodoc_pydantic==2.2.0",
]
torch = [
    "torch>=1.7",
    "wandb",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""

[tool.coverage.run]
plugins = ["covdefaults"]
omit = ["examples"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "testing.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[tool.ruff]
# See all rules here: https://beta.ruff.rs/docs/rules
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # ruff-specific
    "RUF",
]
line-length = 79
extend-ignore = ["Q001"]
target-version = "py38"

[tool.ruff.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.isort]
force-single-line = true
known-first-party = ["deepdrivewe", "test", "testing"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.ruff.per-file-ignores]
"*/__init__.py" = ["F401"]
"*/*_test.py" = ["D10"]

[tool.ruff.lint.pylint]
max-args = 7

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.setuptools.packages.find]
exclude = ["tests*", "testing*"]
namespaces = false
